bugfix: 修复部分不应出现的系统错误信息 #507
