+++ src/backend/ci/core/common/common-api/src/main/kotlin/com/tencent/devops/common/api/pojo/ErrorType.kt
@@ -24,12 +24,9 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.worker.common.exception
+package com.tencent.devops.common.api.pojo
 
-import com.tencent.devops.process.pojo.ErrorType
-
-open class TaskExecuteException(
-    open val errorType: ErrorType,
-    open val errorCode: Int,
-    open val errorMsg: String
-) : Throwable(errorMsg)
+enum class ErrorType(val statusName: String, val visiable: Boolean) {
+    SYSTEM("系统运行错误", true), // 0 平台运行报错
+    USER("业务逻辑错误", true), // 1 任务执行报错
+}
