+++ src/backend/ci/core/plugin/codecc-plugin/worker-plugin-codecc/src/main/kotlin/com/tencent/devops/plugin/worker/task/codecc/util/CodeccParamsHelper.kt
@@ -28,6 +28,8 @@ package com.tencent.devops.plugin.worker.task.codecc.util
 
 import com.fasterxml.jackson.core.type.TypeReference
 import com.tencent.devops.common.api.enums.OSType
+import com.tencent.devops.common.api.exception.TaskExecuteException
+import com.tencent.devops.common.api.pojo.ErrorType
 import com.tencent.devops.common.api.util.JsonUtil
 import com.tencent.devops.common.pipeline.enums.BuildScriptType
 import com.tencent.devops.plugin.codecc.pojo.coverity.CoverityProjectType
@@ -35,12 +37,12 @@ import com.tencent.devops.plugin.codecc.pojo.coverity.ProjectLanguage
 import com.tencent.devops.plugin.worker.pojo.CodeccExecuteConfig
 import com.tencent.devops.plugin.worker.task.codecc.LinuxCodeccConstants
 import com.tencent.devops.plugin.worker.task.codecc.WindowsCodeccConstants
+import com.tencent.devops.process.pojo.AtomErrorCode
 import com.tencent.devops.worker.common.CommonEnv
 import com.tencent.devops.worker.common.api.utils.ThirdPartyAgentBuildInfoUtils
 import com.tencent.devops.worker.common.env.AgentEnv
 import com.tencent.devops.worker.common.env.BuildEnv
 import com.tencent.devops.worker.common.env.BuildType
-import com.tencent.devops.worker.common.exception.UserTaskExecuteException
 import com.tencent.devops.worker.common.logger.LoggerService
 import java.io.File
 
@@ -202,7 +204,11 @@ object CodeccParamsHelper {
         return if (scriptType == BuildScriptType.SHELL) {
             val shareCoverityFile = LinuxCodeccConstants.getCovPyFile()
             if (!shareCoverityFile.exists()) {
-                throw UserTaskExecuteException("The coverity file (${shareCoverityFile.canonicalPath}) is not exist")
+                throw throw TaskExecuteException(
+                    errorType = ErrorType.SYSTEM,
+                    errorCode = AtomErrorCode.USER_TASK_OPERATE_FAIL,
+                    errorMsg = "The coverity file (${shareCoverityFile.canonicalPath}) is not exist"
+                )
             }
             val localCoverityFile = File(codeccWorkspace, shareCoverityFile.name)
             shareCoverityFile.copyTo(localCoverityFile, true)
@@ -216,7 +222,11 @@ object CodeccParamsHelper {
         return if (scriptType == BuildScriptType.SHELL) {
             val shareToolFile = LinuxCodeccConstants.getToolPyFile()
             if (AgentEnv.getOS() != OSType.MAC_OS && !shareToolFile.exists()) {
-                throw UserTaskExecuteException("The mutli tool file (${shareToolFile.canonicalPath}) is not exist")
+                throw throw TaskExecuteException(
+                    errorType = ErrorType.SYSTEM,
+                    errorCode = AtomErrorCode.USER_TASK_OPERATE_FAIL,
+                    errorMsg = "The mutli tool file (${shareToolFile.canonicalPath}) is not exist"
+                )
             }
             val localToolFile = File(codeccWorkspace, shareToolFile.name)
             shareToolFile.copyTo(localToolFile, true)
