+++ src/backend/ci/core/process/api-process/src/main/kotlin/com/tencent/devops/process/engine/pojo/BuildInfo.kt
@@ -28,7 +28,7 @@ package com.tencent.devops.process.engine.pojo
 
 import com.tencent.devops.common.pipeline.enums.BuildStatus
 import com.tencent.devops.common.pipeline.enums.ChannelCode
-import com.tencent.devops.process.pojo.ErrorType
+import com.tencent.devops.common.api.pojo.ErrorType
 
 data class BuildInfo(
     val projectId: String,
