+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/engine/dao/PipelineBuildDao.kt
@@ -33,7 +33,7 @@ import com.tencent.devops.common.pipeline.enums.StartType
 import com.tencent.devops.model.process.Tables.T_PIPELINE_BUILD_HISTORY
 import com.tencent.devops.model.process.tables.records.TPipelineBuildHistoryRecord
 import com.tencent.devops.process.engine.pojo.BuildInfo
-import com.tencent.devops.process.pojo.ErrorType
+import com.tencent.devops.common.api.pojo.ErrorType
 import org.jooq.Condition
 import org.jooq.DSLContext
 import org.jooq.DatePart
