+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/engine/dao/PipelineBuildTaskDao.kt
@@ -34,7 +34,7 @@ import com.tencent.devops.model.process.Tables.T_PIPELINE_BUILD_TASK
 import com.tencent.devops.model.process.tables.TPipelineBuildTask
 import com.tencent.devops.model.process.tables.records.TPipelineBuildTaskRecord
 import com.tencent.devops.process.engine.pojo.PipelineBuildTask
-import com.tencent.devops.process.pojo.ErrorType
+import com.tencent.devops.common.api.pojo.ErrorType
 import org.jooq.DSLContext
 import org.jooq.Result
 import org.slf4j.LoggerFactory
