+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/engine/pojo/event/PipelineBuildFinishEvent.kt
@@ -31,7 +31,7 @@ import com.tencent.devops.common.event.dispatcher.pipeline.mq.MQ
 import com.tencent.devops.common.event.enums.ActionType
 import com.tencent.devops.common.event.pojo.pipeline.IPipelineEvent
 import com.tencent.devops.common.pipeline.enums.BuildStatus
-import com.tencent.devops.process.pojo.ErrorType
+import com.tencent.devops.common.api.pojo.ErrorType
 
 /**
  *
