+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/engine/service/PipelineBuildTaskService.kt
@@ -26,12 +26,12 @@
 
 package com.tencent.devops.process.engine.service
 
+import com.tencent.devops.common.api.pojo.ErrorType
 import com.tencent.devops.common.api.util.JsonUtil
 import com.tencent.devops.common.api.util.timestampmilli
 import com.tencent.devops.common.pipeline.enums.BuildStatus
 import com.tencent.devops.common.pipeline.pojo.element.ElementAdditionalOptions
 import com.tencent.devops.process.engine.dao.PipelineBuildTaskDao
-import com.tencent.devops.process.pojo.ErrorType
 import com.tencent.devops.process.pojo.task.PipelineBuildTaskInfo
 import org.jooq.DSLContext
 import org.springframework.beans.factory.annotation.Autowired
