+++ src/backend/ci/core/worker/worker-api-sdk/src/main/kotlin/com/tencent/devops/worker/common/api/dispatch/CodeccDownloadResourceApi.kt
@@ -27,8 +27,10 @@
 package com.tencent.devops.worker.common.api.dispatch
 
 import com.tencent.devops.common.api.enums.OSType
+import com.tencent.devops.common.api.exception.TaskExecuteException
+import com.tencent.devops.common.api.pojo.ErrorType
+import com.tencent.devops.process.pojo.AtomErrorCode
 import com.tencent.devops.worker.common.api.AbstractBuildResourceApi
-import com.tencent.devops.worker.common.exception.SystemServiceExecuteException
 import okhttp3.Protocol
 import okhttp3.Response
 import org.springframework.http.HttpStatus
@@ -48,7 +50,11 @@ class CodeccDownloadResourceApi : AbstractBuildResourceApi(), CodeccDownloadApi 
         }
         if (!response.isSuccessful) {
             logger.warn("Fail to request($request) with code ${response.code()} , message ${response.message()} and response (${response.body()?.string()})")
-            throw SystemServiceExecuteException("下载Codecc的 $tool 工具失败")
+            throw TaskExecuteException(
+                errorType = ErrorType.SYSTEM,
+                errorCode = AtomErrorCode.SYSTEM_SERVICE_ERROR,
+                errorMsg = "下载Codecc的 $tool 工具失败"
+            )
         }
         return response
     }
@@ -65,7 +71,11 @@ class CodeccDownloadResourceApi : AbstractBuildResourceApi(), CodeccDownloadApi 
         }
         if (!response.isSuccessful) {
             logger.warn("Fail to request($request) with code ${response.code()} , message ${response.message()} and response (${response.body()?.string()})")
-            throw SystemServiceExecuteException("下载codecc的coverity的执行脚本失败")
+            throw TaskExecuteException(
+                errorType = ErrorType.SYSTEM,
+                errorCode = AtomErrorCode.SYSTEM_SERVICE_ERROR,
+                errorMsg = "下载codecc的coverity的执行脚本失败"
+            )
         }
         return response
     }
@@ -83,7 +93,11 @@ class CodeccDownloadResourceApi : AbstractBuildResourceApi(), CodeccDownloadApi 
 
         if (!response.isSuccessful) {
             logger.warn("Fail to request($request) with code ${response.code()} , message ${response.message()} and response (${response.body()?.string()})")
-            throw SystemServiceExecuteException("下载codecc的多工具执行脚本失败")
+            throw TaskExecuteException(
+                errorType = ErrorType.SYSTEM,
+                errorCode = AtomErrorCode.SYSTEM_SERVICE_ERROR,
+                errorMsg = "下载codecc的多工具执行脚本失败"
+            )
         }
         return response
     }
