+++ src/backend/ci/gradle.properties
@@ -40,7 +40,11 @@ MAVEN_REPO_SNAPSHOT_URL=
 MAVEN_REPO_SNAPSHOT_DEPLOY_URL=
 MAVEN_REPO_SNAPSHOT_USERNAME=
 MAVEN_REPO_SNAPSHOT_PASSWORD=
-DB_HOST=127.0.0.1:3306
-DB_USERNAME=devops
-DB_PASSWORD=123456
-DB_PREFIX=devops_ci_
+#DB_HOST=127.0.0.1:3306
+#DB_USERNAME=devops
+#DB_PASSWORD=123456
+#DB_PREFIX=devops_ci_
+DB_HOST=10.125.48.80:3306
+DB_USERNAME=tig4php
+DB_PASSWORD=tig4php
+DB_PREFIX=devops_
